Begin Boat Work
Boat Item and Entity was added, but boats utilize an enum type in BoatEntity.Type that Mixins is not capable of adding to. Other mods replicate the functionality manually, not sure what to do right now. Currently the Coconut Boat item spawns an Oak Boat which functions strictly like an Oak Boat.

Implementation may need to be completely changed to accomplish adding Boats.

Removed TropicaSignEntityRenderer since all it did was extend the SignEntityRenderer.